working on finding friends for g1 and g2 curves
